+++ CHANGELOG.md
@@ -372,3 +372,4 @@
 * Rename `Stack` datastructure to `List`
 * Define `Stack` interface with `push` and `pop` methods
 * Define new type synonymous: `:=:=>`
+* Define experimental `Prepend` verb for `Morphable` ability